+++ pybamm/input/parameters/lithium_ion/positive_electrodes/nmc_ORegan2021/nmc_LGM50_entropic_change_ORegan2021.py
@@ -1,0 +1,38 @@
+from pybamm import exp
+
+
+def nmc_LGM50_entropic_change_ORegan2021(sto):
+    """
+    LG M50 NMC 811 entropic change in open circuit potential (OCP) at a temperature of
+    298.15K as a function of the stochiometry. The fit is taken from [1].
+
+    References
+    ----------
+    .. [1] Kieran O’Regan, Ferran Brosa Planella, W. Dhammika Widanage, and Emma
+    Kendrick. "Thermal-electrochemical parametrisation of a lithium-ion battery:
+    mapping Li concentration and temperature dependencies." Journal of the
+    Electrochemical Society, submitted (2021).
+
+    Parameters
+    ----------
+    sto: :class:`pybamm.Symbol`
+       Electrode stochiometry
+
+    Returns
+    -------
+    :class:`pybamm.Symbol`
+       Entropic change [V.K-1]
+    """
+    a1 = 0.04006
+    a2 = -0.06656
+    b1 = 0.2828
+    b2 = 0.8032
+    c1 = 0.0009855
+    c2 = 0.02179
+
+    dUdT = (
+        a1 * exp(-((sto - b1) ** 2) / c1) + a2 * exp(-((sto - b2) ** 2) / c2)
+    ) / 1000
+    # fit in mV / K
+
+    return dUdT