+++ pybamm/parameters/parameter_sets.py
@@ -207,6 +207,17 @@ Xu2019 = {
     "citation": "Xu2019",
 }
 
+ORegan2021 = {
+    "chemistry": "lithium_ion",
+    "cell": "LGM50_ORegan2021",
+    "negative electrode": "graphite_ORegan2021",
+    "separator": "separator_ORegan2021",
+    "positive electrode": "nmc_ORegan2021",
+    "electrolyte": "lipf6_EC_EMC_3_7_Landesfeind2019",
+    "experiment": "1C_discharge_from_full_ORegan2021",
+    "citation": ["ORegan2021", "Chen2020"],
+}
+
 #
 # Lead-acid
 #